pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "steplight"
version = "0.1.0"
description = "Local-first trace inspector for LLM agents and tool-driven workflows."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
license-files = ["LICENSE"]
authors = [
  { name = "Riccardo Merenda" }
]
keywords = [
  "llm",
  "agents",
  "tracing",
  "observability",
  "tui",
  "cli",
  "debugging",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: Testing",
  "Topic :: Terminals",
]
dependencies = [
  "jinja2>=3.1",
  "pygments>=2.17",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "rich>=13.7",
  "textual>=0.76",
  "typer>=0.12"
]

[project.optional-dependencies]
dev = [
  "build>=1.2",
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
  "twine>=5.1"
]

[project.scripts]
steplight = "steplight.cli.main:app"
slt = "steplight.cli.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -p no:cacheprovider"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["steplight*"]

[tool.setuptools.package-data]
steplight = ["export/templates/*.jinja"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
]
